Exploring Facial Recognition with python scikit-learn
In this project we will use the scikit-learn Olivetti facial data set to explore image processing and facial recognition. The data set is part of the scikit-learn package.  Credit for these images is thanks to AT&T Laboratories Cambridge, MA. The data set consists of 10 different images of 40 subjects for a total of 400 images.